requirements.txt
psycopg2-binary
Pillow==10.0.0
Django==4.2.4
python-dotenv==1.0.0
